+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to `laravel-tracy` will be documented in this file
 
 ## 1.8.5 - 2017-03-08
 
-- minify panel
+- 除了 Terminal Panel之外，其他的 Panel 的 HTML 全部最小化
+- Terminal Panel 增加最小高度
+- Terminal Panel 顯示錯誤訊息
+- Database Panel highlight 重構
 
 ## 1.8.4 - 2017-03-06